requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.11.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
cycler==0.10.0
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.7
django-bootstrap-v5==1.0.11
django-cors-headers==3.14.0
django-on-heroku==1.1.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
Flask==2.2.3
gunicorn==20.1.0
h11==0.12.0
httpie==3.2.1
idna==3.3
importlib-metadata==6.0.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.3.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.3.4
mdurl==0.1.2
multidict==6.0.4
numpy==1.20.1
openapi-codec==1.3.2
outcome==1.1.0
packaging==23.0
Pillow==8.1.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
Pygments==2.14.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2022.7.1
requests==2.28.1
requests-toolbelt==0.10.1
rich==13.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scipy==1.6.2
selenium==4.0.0
simplejson==3.18.3
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
trio==0.19.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.2.3
whitenoise==6.3.0
wsproto==1.0.0
zipp==3.15.0
